Refactor TorqueMeterClient and enhance logging
- Change `_semaphore` to readonly in `TorqueMeterClient`.
- Introduce `isFirst` flag for initial device info reading.
- Improve error handling and logging for better traceability.
- Update `AgingData` and `CheckPointData` with custom `ToString` methods.
- Modify `Station` class to use double array for `CheckingPoints`.
- Refactor `StationAutomationManager` to use `ActivatorUtilities`.
- Integrate Serilog for enhanced logging configuration.
- Adjust frontend components for improved user experience and data handling.

+++ SmartWorkStation.AutomationInspect.App/Services/CheckPointData.cs
@@ -1,3 +1,3 @@
 ﻿namespace SmartWorkStation.AutomationInspect.App.Services;
 
-public record CheckPointData(int Point, int Index,double? ScrewTorque, double? MeterTorque);
+public record CheckPointData(double Point, int Index,double? ScrewTorque, double? MeterTorque);
